Trait salsa::debug::DebugQueryTable

pub trait DebugQueryTable {
    type Key;
    type Value;

    // Required methods
    fn durability(&self, key: Self::Key) -> Durability;
    fn entries<C>(&self) -> C
       where C: FromIterator<TableEntry<Self::Key, Self::Value>>;
}

Additional methods on queries that can be used to “peek into” their current state. These methods are meant for debugging and observing the effects of garbage collection etc.

Required Associated Types

type Key

Key of this query.

type Value

Value of this query.

Required Methods

fn durability(&self, key: Self::Key) -> Durability

Returns a lower bound on the durability for the given key. This is typically the minimum durability of all values that the query accessed, but we may return a lower durability in some cases.

fn entries<C>(&self) -> C
where C: FromIterator<TableEntry<Self::Key, Self::Value>>,

Get the (current) set of the entries in the query table.

Object Safety

This trait is not object safe.