Trait salsa::debug::DebugQueryTable

pub trait DebugQueryTable {
    type Key;
    type Value;
    fn is_constant(&self, key: Self::Key) -> bool;

    fn entries<C>(&self) -> C
    where
        C: FromIterator<TableEntry<Self::Key, Self::Value>>;
}

Additional methods on queries that can be used to "peek into" their current state. These methods are meant for debugging and observing the effects of garbage collection etc.

Associated Types

type Key

Key of this query.

type Value

Value of this query.

Loading content...
Required methods

fn is_constant(&self, key: Self::Key) -> bool

True if salsa thinks that the value for key is a constant, meaning that it can never change, no matter what values the inputs take on from this point.

fn entries<C>(&self) -> C where
C: FromIterator<TableEntry<Self::Key, Self::Value>>,

Get the (current) set of the entries in the query table.

Loading content...