Trait salsa::debug::DebugQueryTable

pub trait DebugQueryTable {
    type Key;
    fn is_constant(&self, key: Self::Key) -> bool;

    fn keys<C>(&self) -> C
    where
        C: FromIterator<Self::Key>;
}

Additional methods on queries that can be used to "peek into" their current state. These methods are meant for debugging and observing the effects of garbage collection etc.

Associated Types

type Key

Key of this query.

Required Methods

fn is_constant(&self, key: Self::Key) -> bool

True if salsa thinks that the value for key is a constant, meaning that it can never change, no matter what values the inputs take on from this point.

fn keys<C>(&self) -> C where
C: FromIterator<Self::Key>,

Get the (current) set of the keys in the query table.