Enum salsa::EventKind

pub enum EventKind {
    DidValidateMemoizedValue {
        database_key: DatabaseKeyIndex,
    },
    WillBlockOn {
        other_runtime_id: RuntimeId,
        database_key: DatabaseKeyIndex,
    },
    WillExecute {
        database_key: DatabaseKeyIndex,
    },
}

An enum identifying the various kinds of events that can occur.

Variants

DidValidateMemoizedValue

Fields

database_key: DatabaseKeyIndex

The database-key for the affected value. Implements Debug.

Occurs when we found that all inputs to a memoized value are up-to-date and hence the value can be re-used without executing the closure.

Executes before the “re-used” value is returned.

WillBlockOn

Fields

other_runtime_id: RuntimeId

The id of the runtime we will block on.

database_key: DatabaseKeyIndex

The database-key for the affected value. Implements Debug.

Indicates that another thread (with id other_runtime_id) is processing the given query (database_key), so we will block until they finish.

Executes after we have registered with the other thread but before they have answered us.

(NB: you can find the id of the current thread via the salsa_runtime)

WillExecute

Fields

database_key: DatabaseKeyIndex

The database-key for the affected value. Implements Debug.

Indicates that the function for this query will be executed. This is either because it has never executed before or because its inputs may be out of date.