Enum salsa::EventKind

pub enum EventKind<DB: Database> {
    DidValidateMemoizedValue {
        database_key: DB::DatabaseKey,
    },
    WillBlockOn {
        other_runtime_id: RuntimeId,
        database_key: DB::DatabaseKey,
    },
    WillChangeInputValue {
        database_key: DB::DatabaseKey,
    },
    WillExecute {
        database_key: DB::DatabaseKey,
    },
}

An enum identifying the various kinds of events that can occur.

Variants

DidValidateMemoizedValue

Occurs when we found that all inputs to a memoized value are up-to-date and hence the value can be re-used without executing the closure.

Executes before the "re-used" value is returned.

Fields of DidValidateMemoizedValue

database_key: DB::DatabaseKey

The database-key for the affected value. Implements Debug.

WillBlockOn

Indicates that another thread (with id other_runtime_id) is processing the given query (database_key), so we will block until they finish.

Executes after we have registered with the other thread but before they have answered us.

(NB: you can find the id of the current thread via the salsa_runtime)

Fields of WillBlockOn

other_runtime_id: RuntimeId

The id of the runtime we will block on.

database_key: DB::DatabaseKey

The database-key for the affected value. Implements Debug.

WillChangeInputValue

Indicates that the input value will change after this callback, e.g. due to a call to set.

Fields of WillChangeInputValue

database_key: DB::DatabaseKey

The database-key for the affected value. Implements Debug.

WillExecute

Indicates that the function for this query will be executed. This is either because it has never executed before or because its inputs may be out of date.

Fields of WillExecute

database_key: DB::DatabaseKey

The database-key for the affected value. Implements Debug.